Small fixxxxx

+++ Patches/SCP914ProcessItems.cs
@@ -100,47 +100,42 @@ namespace Better914.Patches
             int v = 0;
             if (knobState == -2)
             {
-                v = GetRandomItem(
-                    new Dictionary<int, int> {
-			{ 100 - ( PluginConfig.Cfg.Level_4Chance + PluginConfig.Cfg.Level_3Chance ), -2 },
-			{ PluginConfig.Cfg.Level_3Chance, -3 },
-			{ PluginConfig.Cfg.Level_4Chance, -4 }
-		} );
+                v = GetRandomItem( new Dictionary<float, int> {
+                        { 100 - ( PluginConfig.Cfg.Level_4Chance + PluginConfig.Cfg.Level_3Chance ), -2 },
+                        { PluginConfig.Cfg.Level_3Chance, -3 },
+                        { PluginConfig.Cfg.Level_4Chance, -4 }
+                    });
             }
             else if (knobState == -1)
             {
-                v = GetRandomItem(
-                    new Dictionary<int, int> {
-			{ PluginConfig.Cfg.SameItemChance, 10 },
-			{ 100 - ( PluginConfig.Cfg.Level_2Chance ), -1 },
-			{ PluginConfig.Cfg.Level_2Chance, -2 }
-                    } );
+                v = GetRandomItem( new Dictionary<float, int> {
+                        { PluginConfig.Cfg.SameItemChance, 10 },
+                        { 100 - ( PluginConfig.Cfg.Level_2Chance ), -1 },
+                        { PluginConfig.Cfg.Level_2Chance, -2 }
+                    });
             }
             else if (knobState == 0)
             {
-		v = GetRandomItem(
-                    new Dictionary<int, int> {
-			{ PluginConfig.Cfg.SameItemChance, 10 },
-			{ 100 - (PluginConfig.Cfg.SameItemChance), 0 }
-            	} );
+                v = GetRandomItem( new Dictionary<float, int> {
+                        { PluginConfig.Cfg.SameItemChance, 10 },
+                        { 100 - (PluginConfig.Cfg.SameItemChance), 0 }
+                    });
             }
             else if (knobState == 1)
             {
-		v = GetRandomItem(
-                    new Dictionary<int, int> {
-			{ PluginConfig.Cfg.SameItemChance, 10 },
-			{ 100 - ( PluginConfig.Cfg.Level2Chance ), 1 },
-			{ PluginConfig.Cfg.Level2Chance, 2 }
-            	} );
+                v = GetRandomItem( new Dictionary<float, int> {
+                        { PluginConfig.Cfg.SameItemChance, 10 },
+                        { 100 - ( PluginConfig.Cfg.Level2Chance ), 1 },
+                        { PluginConfig.Cfg.Level2Chance, 2 }
+                    });
             }
             else if (knobState == 2)
             {
-                v = GetRandomItem(
-                    new Dictionary<int, int> {
-			{ 100 - ( PluginConfig.Cfg.Level4Chance + PluginConfig.Cfg.Level3Chance ), 2 },
-			{ PluginConfig.Cfg.Level3Chance, 3 },
-			{ PluginConfig.Cfg.Level4Chance, 4 }
-		} );
+                v = GetRandomItem( new Dictionary<float, int> {
+                        { 100 - ( PluginConfig.Cfg.Level4Chance + PluginConfig.Cfg.Level3Chance ), 2 },
+                        { PluginConfig.Cfg.Level3Chance, 3 },
+                        { PluginConfig.Cfg.Level4Chance, 4 }
+                    });
             }
             
             if (v == 10) return item; //10 = the same item
@@ -331,18 +326,18 @@ namespace Better914.Patches
             return (number < min) ? min : (number > max) ? max : number;
         }
 
-	public static int GetRandomItem(Dictionary<int, int> chances)//int, int = chance, level
-	{
-		var r = Random.Range(0, 10);
-		r *= 10;
-		int sum = 0;
-		for(int i = 0; i < chances.Count; i++)
-		{
-			if (r <= chances[i] + sum) return i;
-			else sum += chances[i];
-		}
-		return 10; //if total chances are < 100% this can occur, the item wont be upgraded
-	}
+    public static int GetRandomItem(Dictionary<float, int> chances)//int, int = chance, level
+    {
+        var r = Random.Range(0, 10);
+        r *= 10;
+        int sum = 0;
+        for(int i = 0; i < chances.Count; i++)
+        {
+            if (r <= chances[i] + sum) return i;
+            else sum += chances[i];
+        }
+        return 10; //if total chances are < 100% this can occur, the item wont be upgraded
+    }
 	
 	public static bool CheckPercent(float chance)
 	{
